requirements.txt
backoff==2.2.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
Deprecated==1.2.14
dnspython==2.3.0
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.0
idna==3.4
importlib-metadata==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.1
opentelemetry-api==1.18.0
opentelemetry-distro==0.39b0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-exporter-otlp-proto-http==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-instrumentation-flask==0.39b0
opentelemetry-instrumentation-requests==0.39b0
opentelemetry-instrumentation-wsgi==0.39b0
opentelemetry-proto==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
opentelemetry-util-http==0.39b0
pandas==2.0.3
protobuf==4.23.4
psycopg2-binary==2.9.6
pymongo==4.4.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
uWSGI==2.0.21
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.0
